+++ Nernel/Source/System/PIT.h
@@ -10,5 +10,5 @@ void PIT_TestAPICFrequency(APIC* pAPIC);
 bool_t PIT_TestingAPICFrequency();
 
 extern uint16_t gPIT_CycleHertz;
-extern uint32_t gPIT_MillisecondsElapsed;
+extern uint16_t gPIT_MillisecondsElapsed;
 extern uint32_t gPIT_SecondsElapsed;
